requirements.txt
fastapi
uvicorn
motor
pymongo
beautifulsoup4
requests
aioredis
python-dotenv
pydantic
